Add supplier fixture
