+++ apps/e2e/helpers/fixtures.ts
@@ -1,11 +1,12 @@
 import { baseURL } from "@/integration/constants";
 import test from "@playwright/test";
-import { upsertBook, upsertCustomer } from "./cr-sqlite";
+import { addBooksToCustomer, associatePublisher, upsertBook, upsertCustomer, upsertSupplier } from "./cr-sqlite";
 import { getDbHandle } from "./db";
 
 type OrderTestFixture = {
 	customer: { id: number; fullname: string; email: string; displayId: string };
 	books: { isbn: string; authors: string; title: string; publisher: string; price: number }[];
+	supplier: { id: number; name: string; email: string };
 };
 
 export const testOrders = test.extend<OrderTestFixture>({
@@ -35,5 +36,27 @@ export const testOrders = test.extend<OrderTestFixture>({
 		await dbHandle.evaluate(upsertBook, books[1]);
 
 		await use(books);
+	},
+	supplier: async ({ page }, use) => {
+		await page.goto(baseURL);
+
+		const customer = { id: 1, fullname: "John Doe", email: "john@gmail.com", displayId: "1" };
+		const supplier = { id: 1, name: "Sup1", email: "sup1@gmail.com" };
+
+		const dbHandle = await getDbHandle(page);
+
+		// dbHandler
+		await dbHandle.evaluate(upsertCustomer, customer);
+		await dbHandle.evaluate(addBooksToCustomer, { customerId: 1, bookIsbns: ["1234"] });
+		await dbHandle.evaluate(addBooksToCustomer, { customerId: 1, bookIsbns: ["1234", "4321"] });
+
+		await dbHandle.evaluate(upsertSupplier, supplier);
+
+		await dbHandle.evaluate(associatePublisher, { supplierId: 1, publisherId: "pub1" });
+		await dbHandle.evaluate(upsertBook, { isbn: "1234", title: "Book1", authors: "Author1", publisher: "pub1", price: 12 });
+
+		await dbHandle.evaluate(upsertBook, { isbn: "4321", title: "Book2", authors: "Author2", publisher: "pub1", price: 13 });
+
+		await use(supplier);
 	}
 });
